BasicMechanics
Walls are in. Added a interactable option; Player has a hitbox that is seperate from the wall personal collider. Added humans as interactables. Can bring humans to the ufo spot and leave them. Can only carry one person at a time. No visuals to show the player whats going on.