Enemy AI; Listen And Audio
Listener is a script that only has a function that returns a bool if the player is in the radius and can hear them. Vision shoots out a raycast that hits walls and the player. If it hits the player the enemy can see them. Enemy movement has not been implemented.