Exclamation & Question
A question mark appears on the side of the screen if an enemy has heard you.
A exclamation mark appears on the side of the screen if the enemy is chasing you. There is an issue where the enemy is on screen and will still have these symbols. They wont dissapear even after adding a test for the enemies location within the player range